[FRONT] Changed card component
