+++ ram-back/package.json
@@ -21,6 +21,7 @@
     "@aws-sdk/credential-providers": "^3.303.0",
     "@types/bcrypt": "^5.0.0",
     "@types/minio": "^7.0.18",
+    "@types/react-datepicker": "^4.10.0",
     "bcrypt": "^5.1.0",
     "body-parser": "^1.20.2",
     "cors": "^2.8.5",
@@ -42,6 +43,7 @@
     "pg": "^8.10.0",
     "prettier": "^2.8.4",
     "prettier-eslint": "^15.0.1",
+    "react-datepicker": "^4.11.0",
     "reflect-metadata": "^0.1.13",
     "sinon": "^15.0.3",
     "ts-node": "^10.9.1",
